Modularize Arcanist settings, let users pick between ambiguous interpretations of `arc browse ...`
Summary:
Ref T10895. This modularizes Arcanist settings, but doesn't do anything with them yet.

When users type `arc browse X`, and only provide one "X", and there are several ways to interpret it, prompt them to choose one.

Test Plan: Created a file named `T234`, ran `arc browse T234`, was prompted to interpret it as an object or a path.

Reviewers: chad

Reviewed By: chad

Maniphest Tasks: T10895

Differential Revision: https://secure.phabricator.com/D16928