Add support for enforcing a template for revert plans (#105)
When enabled, this change prompts the developer for whether a change is covered by a feature flag or not, and generates an initial revert plan for the diff appropriately.  We introduce two new `.arcconfig` settings:
1. `differential.revert_plan_check_paths`: A regular expression describing paths for which the revert plan templating should be enabled.  If this setting is absent, the feature is disabled.
2. `differential.revert_plan_ommitted_reasons`: An array of strings indicating possible reasons why a diff may not be covered by a feature flag.

The logic is mostly straightforward.  The only slightly-intrusive change is that since the revert plan templating should only apply to diffs modifying certain paths, we had to refactor the logic to call `generateChanges()` earlier when the `differential.revert_plan_check_paths` is present.