[Wilds] Stop writing temporary files for linter tests
Summary:
Ref T13209. See some discussion in T13209#241713.

There's a bug here where `0644` (a numeric literal written in octal) should be `'0644'` (a string literal). This caused the `TempFile` to fail to `unlink()` on Windows.

But with `ArcanistWorkingCopyPath` we don't have to write an actual file to disk at all, so just don't. Maybe we'll start doing this some of the time later on (e.g., to make it easier to test third-party linters which do not have a "read from stdin" mode), but ideally we shouldn't need to actually write to disk in order to test that linters work, at least in most cases.

Also fix the octal bug itself.

Test Plan: Under Windows, more tests pass and there's no more `unlink()` permissions error.

Reviewers: amckinley

Reviewed By: amckinley

Maniphest Tasks: T13209

Differential Revision: https://secure.phabricator.com/D19728