[Wilds] Make "arc anoid" system requirements more accurate
Summary:
Fixes T8693. Ref T13098. On a 30x15 terminal, the we can only fit "Score: X/12 * Deaths: Y" on the top line if both `X` and `Y` are less than 10, so they can render with a single character.

As soon as the player breaks more than 9 blocks or dies more than 9 times, we need an extra character to render the score. This causes an off-screen write to curses and crashes.

Raise the minimum requirement to 32 columns so we can render "12/12" and up to "99" deaths. Then, change the display logic to show "99" if you die more than 99 times.

(At this resolution we always generate a board with 12 blocks, even if the terminal is very very tall, so we don't need to deal with a case where the "Score" might read "101/200".)

Test Plan:
- Beat the game on a 32x15 terminal.
- Changed logic to award me 1000 deaths per actual death.
- Died on a 32x15 terminal.

Reviewers: amckinley

Reviewed By: amckinley

Maniphest Tasks: T13098, T8693

Differential Revision: https://secure.phabricator.com/D20085