Inform user when configuration is incorrect for Configurable test engines (#121)
If `unit.golang.command` or `unig.pytest.command` was missing for `ConfigurableGolandTestEngine` and `ConfigurablePytestTestEngine` strange exception was thrown with no clear explanation how to fix that.

This patch adds validation of mentioned config options which are required for test engines. Exception is thrown in following conditions:
1. If option is missing.
2. If option does not have two `%s` placeholders which are used as paths for `JNUIT_XML` and `COVERAGE_XML` variables.

Both two classes were blacklisted on `xhpast` linter. The linter indicated errors of unsafe `csprintf` use. This is technical dept and refactoring is necessary to fix the issue. But that would require changing config files on multiple repositories.

Resolves #48