Make GOVET linter a no-op. (#125)
* linter: new "noop" linter

This noop linter is a linter class that does nothing at all. It is
intended to be used when a previously working linter doesn't work
anymore and should be disabled instead of fixed.

Derived linters should only need to implement
`getLinterConfigurationName`.

* linter: make govet a noop linter

`go tool vet` was finally removed in Go 1.12. The old invocation allowed
files to be individually passed but the new one wants import paths to
packages. This is a major change to how `ArcanistGoVetLinter` needs to
work.

For the time being, disable `govet` linting by turning it into a noop
linter.