Allow long lines if the length is due to URLs
It is at times desirable to include URLs in source files (for example
liking to an issue tracker or online documentation as a justification
for particular piece of code).  Sadly, website addresses may easily
end up line-length limit configured in a project.

User is left with two unappealing alternatives:
1) ignore lint warning which makes the build process more noisy or
2) wrap the URL over multiple lines making it inconvenient to follow.

This commit introduces a text.allow-long-lines-with-urls option,
enabled by default, which silences the "line too long" warning if the
URL is the main factor contributing to the length.  This gives
a third, most appealing alternative of simply isolating URL to be on
its own line.